requirements.txt
altgraph==0.17.4
kaleido==0.2.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
plotly==5.23.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
setuptools==72.1.0
six==1.16.0
tenacity==9.0.0
ttkbootstrap==1.10.1
tzdata==2024.1
